+++ include/ML/Engine/EngineEvents.hpp
@@ -37,7 +37,7 @@ namespace ml
 			MAX_ENGINE_EVENT
 		};
 
-		static_assert(MAX_ENGINE_EVENT < (MIN_ENGINE_EVENT + Event::MAX_LIBRARY_EVENTS),
+		static_assert(MAX_ENGINE_EVENT < (MIN_ENGINE_EVENT + (int32_t)Event::MAX_LIBRARY_EVENTS),
 			"too many library event types specified in " __FILE__
 		);
 	};
